changed int to Long

+++ src/main/java/org/springframework/samples/travel/HotelsController.java
@@ -48,7 +48,7 @@ public class HotelsController {
     */
 
 	@RequestMapping(value = "/hotels/booking", method = RequestMethod.GET)
-	public String bookHotel(@RequestParam int hotelId,Principal currentUser) {
+	public String bookHotel(@RequestParam Long hotelId,Principal currentUser) {
 		model.addAttribute(bookingService.findHotelById(hotelId));
 		System.out.println("Hotel id is :" + hotelId);
 		System.out.println("User logged in is "+ currentUser.getName());
